Add files via upload
This system is based on a concept to maintain and generate all the four wheeler’s record of a showroom. From this system, the user can easily maintain each and every four wheeler’s record. The user can add, edit, view and remove car details. Cars can be searched according to their used years or price and distance traveled which makes easy to search for the customers. The whole project is developed in JAVA, different variables and strings have been used for the development of this project.